Spotify: CategoryCollectionViewCell redesigned
