+++ Spotify/Presentation layer/Search/CategoryCollectionViewCell.swift
@@ -1,0 +1,79 @@
+//
+//  GenreCollectionViewCell.swift
+//  Spotify
+//
+//  Created by Natalia Shchipakina on 28.08.2023.
+//
+
+import UIKit
+import Kingfisher
+
+class CategoryCollectionViewCell: UICollectionViewCell {
+    static let identifier = "GenreCollectionViewCell"
+    
+    //    MARK: - UI
+    
+    private let imageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.contentMode = .scaleAspectFill
+        imageView.tintColor = .gray
+        return imageView
+    }()
+    
+    private let label: UILabel = {
+        let label = UILabel()
+        label.textColor = .white
+        label.font = .systemFont(ofSize: 22, weight: .semibold)
+        label.numberOfLines = 0
+        return label
+    }()
+    
+    //    MARK: Init
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupUI()
+        setupConstraints()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        label.text = nil
+        imageView.image = nil
+    }
+    
+    private func setupUI() {
+        backgroundColor = .clear
+        contentView.layer.cornerRadius = 8
+        contentView.layer.masksToBounds = true
+        contentView.addSubviews(imageView, label)
+    }
+    
+    private func setupConstraints() {
+        label.translatesAutoresizingMaskIntoConstraints = false
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        
+        NSLayoutConstraint.activate([
+            label.topAnchor.constraint(greaterThanOrEqualTo: contentView.topAnchor, constant: 10),
+            label.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 10),
+            label.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -10),
+            label.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -6)
+        ])
+        
+        NSLayoutConstraint.activate([
+            imageView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            imageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
+        ])
+    }
+    
+    func configure(with title: String, imageURL: URL) {
+        label.text = title
+        imageView.kf.setImage(with: imageURL)
+    }
+}
